+++ src/main/java/me/lemonboi439/archeryRevamped/item/ShockwaveArrowItem.java
@@ -1,10 +1,15 @@
 package me.lemonboi439.archeryRevamped.item;
 
-import net.minecraft.item.ArrowItem;
+import me.lemonboi439.archeryRevamped.arrow.ArrowType;
 import net.minecraft.item.Item;
 
-public final class ShockwaveArrowItem extends ArrowItem {
+public final class ShockwaveArrowItem extends SpecialArrowItem {
     public ShockwaveArrowItem(Item.Settings settings) {
         super(settings);
+    }
+
+    @Override
+    protected ArrowType arrowType() {
+        return ArrowType.SHOCKWAVE;
     }
 }